adu/requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
python-socketio==5.8.0
gunicorn==22.0.0
celery==5.3.4
redis==5.0.1
psycopg2-binary
vertica-python
polars>=0.20.0
pandera
pandas
pyarrow>=14.0.0
uvloop>=0.19.0
aiofiles>=23.2.1
asyncio-throttle>=1.0.2
psutil>=5.9.0
duckdb>=0.9.0